Create guessing_game.py
The following code was written by myself and is designed to meet the exceeds expectations requirement for project 1 on the tech degree python track. The game will ask the player if they would like to play, give them a set of parameters in which to enter, and once finished ask if they would like to play again or tell them to have a great day. 
I look forward to any and all feedback I can get on this code and wish to utilize said feedback to build a better foundation in my Python knowledge. 
Thank you for stopping by!